stop raycasting, working on minimap

+++ includes/map.h
@@ -6,7 +6,7 @@
 /*   By: glaguyon <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/14 14:29:41 by glaguyon          #+#    #+#             */
-/*   Updated: 2024/06/21 16:41:13 by glaguyon         ###   ########.fr       */
+/*   Updated: 2024/06/21 18:12:23 by glaguyon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,8 +43,8 @@ typedef struct s_map
 	t_img		s;
 	t_img		w;
 	t_img		e;
-	uint64_t	floor;
-	uint64_t	ceil;
+	uint32_t	floor;
+	uint32_t	ceil;
 	t_player	player;
 }	t_map;
